+++ lib/src/platforms/android/file_saver_android.dart
@@ -21,9 +21,9 @@ class FileSaverAndroid extends FileSaverPlatform {
   /// Native FileSaver instance
   late final bindings.FileSaver _fileSaver;
 
-  @override
-  void dispose() {
-    _fileSaver.release();
+  /// Registers this class as the default instance of [FileSaverPlatform].
+  static void registerWith() {
+    FileSaverPlatform.instance = FileSaverAndroid();
   }
 
   @override